Added movement tokens for Alpha Strike in play mode
This adds a small hex token next to pilot skill with a <?>. Clicking on it will show the appropriate movement tokens for the unit. Selecting the movement type will now show the letter and coloring of the movement token in the small hex.

TTM is updated to reflect movement. It should respect JMPS# and JMPW#.

Movement affects to firing should be respected as well. Standstill reduces by 1, jumping increases by 2, jumping jack SPA is considered as well as BA having no jump penalty.