Initial commit of round damage and heat staging
Dots and heat will now indicate the future state in teal. When a unit has round changes staged, it will show an "Apply" button to commit the changes immediately.

There is also now a "End Round" button in the main menu that will apply all the changes for all the units.